+++ apps/api/src/app/portfolio/interfaces/portfolio-position-detail.interface.ts
@@ -19,6 +19,7 @@ export interface PortfolioPositionDetail {
   quantity: number;
   symbol: string;
   transactionCount: number;
+  value: number;
 }
 
 export interface HistoricalDataContainer {