+++ apps/api/src/app/portfolio/portfolio.controller.ts
@@ -370,7 +370,8 @@ export class PortfolioController {
           'grossPerformance',
           'investment',
           'netPerformance',
-          'quantity'
+          'quantity',
+          'value'
         ]);
       }